+++ src/aux-server/aux-web/shared/scene/decorators/BotShapeDecorator.ts
@@ -8,6 +8,12 @@ import {
     getBotSubShape,
     BotSubShape,
     calculateNumericalTagValue,
+    getBotOrientationMode,
+    getBotAnchorPoint,
+    BotOrientationMode,
+    BotAnchorPoint,
+    calculateStringTagValue,
+    hasValue,
 } from '@casual-simulation/aux-common';
 import {
     Mesh,
@@ -22,6 +28,12 @@ import {
     Box3,
     Scene,
     Object3D,
+    Vector2,
+    Matrix4,
+    Euler,
+    AnimationMixer,
+    SkinnedMesh,
+    AnimationAction,
 } from 'three';
 import {
     createCube,
@@ -38,6 +50,8 @@ import { IMeshDecorator } from './IMeshDecorator';
 import { ArgEvent } from '@casual-simulation/aux-common/Events';
 import { GLTF } from 'three/examples/jsm/loaders/GLTFLoader';
 import { getGLTFPool } from '../GLTFHelpers';
+import { HtmlMixer, HtmlMixerHelpers } from '../HtmlMixer';
+import { Game } from '../Game';
 
 const gltfPool = getGLTFPool('main');
 
@@ -47,15 +61,30 @@ export class BotShapeDecorator extends AuxBot3DDecoratorBase
     private _subShape: BotSubShape = null;
     private _gltfVersion: number = null;
     private _address: string = null;
+    private _animation: any = null;
     private _canHaveStroke = false;
+    private _animationMixer: AnimationMixer;
+    private _animClips: AnimationAction[];
+    private _animClipMap: Map<string, AnimationAction>;
+
+    /**
+     * The 3d plane object used to display an iframe.
+     */
+    private _iframe: HtmlMixer.Plane;
+
+    private _game: Game;
 
     container: Group;
-    mesh: Mesh | Sprite;
+    mesh: Mesh;
     collider: Object3D;
     scene: Scene;
 
     get allowModifications() {
-        return this._subShape === null;
+        return this._subShape === null && this._shape !== 'iframe';
+    }
+
+    get allowMaterialModifications() {
+        return this._subShape === null && this._shape !== 'iframe';
     }
 
     /**
@@ -65,15 +94,19 @@ export class BotShapeDecorator extends AuxBot3DDecoratorBase
 
     onMeshUpdated: ArgEvent<IMeshDecorator> = new ArgEvent<IMeshDecorator>();
 
-    constructor(bot3D: AuxBot3D) {
+    constructor(bot3D: AuxBot3D, game: Game) {
         super(bot3D);
 
+        this._game = game;
         this._rebuildShape('cube', null, null, null);
     }
 
-    // frameUpdate?(calc: BotCalculationContext): void {
-
-    // }
+    frameUpdate() {
+        if (this._game && this._animationMixer) {
+            this._animationMixer.update(this._game.getTime().deltaTime);
+            this.scene.updateMatrixWorld(true);
+        }
+    }
 
     botUpdated(calc: BotCalculationContext): void {
         const shape = getBotShape(calc, this.bot3D.bot);
@@ -89,12 +122,19 @@ export class BotShapeDecorator extends AuxBot3DDecoratorBase
             'auxGLTFVersion',
             2
         );
+        const animation = calculateBotValue(
+            calc,
+            this.bot3D.bot,
+            'auxFormAnimation'
+        );
         if (this._needsUpdate(shape, subShape, address, version)) {
             this._rebuildShape(shape, subShape, address, version);
         }
 
         this._updateColor(calc);
         this._updateStroke(calc);
+        this._updateAddress(calc, address);
+        this._updateAnimation(animation);
     }
 
     private _needsUpdate(
@@ -112,7 +152,7 @@ export class BotShapeDecorator extends AuxBot3DDecoratorBase
     }
 
     private _updateStroke(calc: BotCalculationContext) {
-        if (!this._canHaveStroke) {
+        if (!this._canHaveStroke || !this.mesh) {
             return;
         }
 
@@ -130,11 +170,11 @@ export class BotShapeDecorator extends AuxBot3DDecoratorBase
         const hasStroke = typeof strokeColorValue !== 'undefined';
         if (hasStroke && !this.stroke) {
             this.stroke = createStroke();
-            this.container.add(this.stroke);
+            this.mesh.add(this.stroke);
         } else if (!hasStroke) {
             if (this.stroke) {
                 disposeMesh(this.stroke);
-                this.container.remove(this.stroke);
+                this.mesh.remove(this.stroke);
 
                 this.stroke = null;
             }
@@ -158,6 +198,65 @@ export class BotShapeDecorator extends AuxBot3DDecoratorBase
         }
     }
 
+    private _updateAddress(calc: BotCalculationContext, address: string) {
+        if (this._address === address) {
+            return;
+        }
+        this._address = address;
+        if (this._iframe) {
+            if (this._subShape === 'src') {
+                this._updateIframeSrc();
+            } else if (this._subShape === 'html') {
+                this._updateIframeHtml();
+            } else {
+                this._updateIframeHtml();
+            }
+        }
+    }
+
+    private _updateAnimation(animation: any, forceUpdate: boolean = false) {
+        if (this._animation === animation && !forceUpdate) {
+            return;
+        }
+        if (!this._animationMixer) {
+            return;
+        }
+        this._animation = animation;
+        this._animationMixer.stopAllAction();
+
+        const noAnimation = animation === false;
+        if (noAnimation) {
+        } else if (hasValue(this._animation)) {
+            let playing = false;
+            if (typeof this._animation === 'number') {
+                const index = Math.floor(this._animation);
+                if (index >= 0 && index < this._animClips.length) {
+                    this._animClips[index].play();
+                    playing = true;
+                }
+            }
+
+            if (!playing) {
+                const name = this._animation.toString();
+                const clip = this._animClipMap.get(name);
+                if (clip) {
+                    clip.play();
+                    playing = true;
+                }
+            }
+        } else {
+            this._animClips[0].play();
+        }
+    }
+
+    private _updateIframeHtml() {
+        HtmlMixerHelpers.setIframeHtml(this._iframe, this._address);
+    }
+
+    private _updateIframeSrc() {
+        HtmlMixerHelpers.setIframeSrc(this._iframe, this._address);
+    }
+
     dispose(): void {
         const index = this.bot3D.colliders.indexOf(this.collider);
         if (index >= 0) {
@@ -168,13 +267,19 @@ export class BotShapeDecorator extends AuxBot3DDecoratorBase
         disposeMesh(this.mesh);
         disposeMesh(this.stroke);
         disposeObject3D(this.collider);
+        if (this._iframe) {
+            this.container.remove(this._iframe.object3d);
+            disposeObject3D(this._iframe.object3d);
+        }
         disposeScene(this.scene);
 
+        this._animationMixer = null;
         this.mesh = null;
         this.collider = null;
         this.container = null;
         this.scene = null;
         this.stroke = null;
+        this._iframe = null;
     }
 
     private _updateColor(calc: BotCalculationContext) {
@@ -206,7 +311,6 @@ export class BotShapeDecorator extends AuxBot3DDecoratorBase
 
         // Container
         this.container = new Group();
-        this.container.position.set(0, 0.5, 0);
         this.bot3D.display.add(this.container);
 
         if (this._shape === 'cube') {
@@ -221,9 +325,53 @@ export class BotShapeDecorator extends AuxBot3DDecoratorBase
             } else {
                 this._createCube();
             }
+        } else if (this._shape === 'iframe') {
+            if (this._subShape === 'src') {
+                this._createSrcIframe();
+            } else if (this._subShape === 'html') {
+                this._createHtmlIframe();
+            } else {
+                this._createHtmlIframe();
+            }
         }
 
         this.onMeshUpdated.invoke(this);
+    }
+
+    private _createSrcIframe() {
+        if (this._createIframe()) {
+            this._updateIframeSrc();
+        }
+    }
+
+    private _createHtmlIframe() {
+        if (this._createIframe()) {
+            this._updateIframeHtml();
+        }
+    }
+
+    private _createIframe() {
+        if (!this._game) {
+            return false;
+        }
+        const mixerContext = this._game.getHtmlMixerContext();
+        const domElement = HtmlMixerHelpers.createIframeDomElement(
+            'about:blank'
+        );
+
+        this._iframe = new HtmlMixer.Plane(mixerContext, domElement, {
+            elementW: 768,
+            planeW: 1,
+            planeH: 1,
+        });
+
+        this.container.add(this._iframe.object3d);
+
+        this._createCube();
+        this.mesh.scale.set(1, 0.01, 0.05);
+        this.mesh.position.set(0, -0.5, 0);
+
+        return true;
     }
 
     private _createGltf() {
@@ -270,6 +418,22 @@ export class BotShapeDecorator extends AuxBot3DDecoratorBase
         setColor(collider, 'clear');
         this.container.add(this.collider);
         this.bot3D.colliders.push(this.collider);
+
+        // Animations
+        if (gltf.animations.length > 0) {
+            this._animationMixer = new AnimationMixer(this.scene);
+            let clipMap = new Map<string, AnimationAction>();
+            let clips = [] as AnimationAction[];
+            for (let anim of gltf.animations) {
+                const action = this._animationMixer.clipAction(anim);
+                clips.push(action);
+                clipMap.set(anim.name, action);
+            }
+
+            this._animClips = clips;
+            this._animClipMap = clipMap;
+            this._updateAnimation(null, true);
+        }
 
         this.bot3D.updateMatrixWorld(true);
     }