+++ src/aux-server/aux-web/shared/scene/decorators/DimensionPositionDecorator.ts
@@ -13,8 +13,13 @@ import {
     cacheFunction,
     calculateBooleanTagValue,
     isBotStackable,
+    getBotOrientationMode,
+    getBotAnchorPoint,
+    BotAnchorPoint,
+    BotOrientationMode,
+    getBotIndex,
 } from '@casual-simulation/aux-common';
-import { Vector3, Quaternion, Euler, Vector2 } from 'three';
+import { Vector3, Quaternion, Euler, Vector2, Object3D } from 'three';
 import { calculateGridTileLocalCenter } from '../grid/Grid';
 import { realPosToGridPos, Axial, posToKey } from '../hex';
 import { BuilderGroup3D } from '../BuilderGroup3D';
@@ -38,9 +43,13 @@ export class DimensionPositionDecorator extends AuxBot3DDecoratorBase {
     private _atPosition: boolean;
     private _atRotation: boolean;
     private _lastPos: { x: number; y: number; z: number };
+    private _lastSortOrder: number;
     private _nextPos: Vector3;
     private _nextRot: { x: number; y: number; z: number };
     private _lastHeight: number;
+    private _orientationMode: BotOrientationMode;
+    private _anchorPoint: BotAnchorPoint;
+    private _rotationObj: Object3D;
 
     constructor(
         bot3D: AuxBot3D,
@@ -51,6 +60,32 @@ export class DimensionPositionDecorator extends AuxBot3DDecoratorBase {
     }
 
     botUpdated(calc: BotCalculationContext): void {
+        const nextOrientationMode = getBotOrientationMode(calc, this.bot3D.bot);
+        const nextAnchorPoint = getBotAnchorPoint(calc, this.bot3D.bot);
+
+        if (nextOrientationMode !== this._orientationMode) {
+            this.bot3D.display.rotation.set(0, 0, 0);
+        }
+        this._orientationMode = nextOrientationMode;
+        this._anchorPoint = nextAnchorPoint;
+
+        // Update the offsets for the center vs bottom
+        // anchor positions so that the bot is above the grid and
+        // not in it
+        if (this._anchorPoint === 'center') {
+            this.bot3D.display.position.set(0, 0, 0);
+            if (this._rotationObj) {
+                this._rotationObj.rotation.set(0, 0, 0);
+            }
+            this._rotationObj = this.bot3D.display;
+        } else {
+            this.bot3D.display.position.set(0, 0.5, 0);
+            if (this._rotationObj) {
+                this._rotationObj.rotation.set(0, 0, 0);
+            }
+            this._rotationObj = this.bot3D.container;
+        }
+
         const userDimension = this.bot3D.dimension;
         if (userDimension) {
             const scale = this.bot3D.gridScale;
@@ -65,26 +100,53 @@ export class DimensionPositionDecorator extends AuxBot3DDecoratorBase {
                 this.bot3D.dimension,
                 scale
             );
+            const currentSortOrder = getBotIndex(
+                calc,
+                this.bot3D.bot,
+                this.bot3D.dimension
+            );
             this._nextPos = calculateObjectPositionInGrid(
                 calc,
                 this.bot3D,
                 scale
             );
 
-            if (
-                this._positionUpdated(currentGridPos) ||
-                this._heightUpdated(currentHeight)
-            ) {
-                const objectsAtPosition = objectsAtDimensionGridPosition(
-                    calc,
-                    this.bot3D.dimension,
-                    this._lastPos || currentGridPos
-                );
-                this.bot3D.dimensionGroup.simulation3D.ensureUpdate(
-                    objectsAtPosition.map(f => f.id)
+            // Offset the container so that the bot still
+            // is above the grid
+            if (this._anchorPoint === 'center') {
+                this._nextPos.add(
+                    new Vector3(0, 0.5, 0).multiplyScalar(
+                        this.bot3D.container.scale.x
+                    )
                 );
             }
+
+            if (
+                this._positionUpdated(currentGridPos, currentSortOrder) ||
+                this._heightUpdated(currentHeight)
+            ) {
+                let ids = [] as string[];
+                if (this._lastPos) {
+                    const objectsAtLastPosition = objectsAtDimensionGridPosition(
+                        calc,
+                        this.bot3D.dimension,
+                        this._lastPos
+                    );
+                    ids.push(...objectsAtLastPosition.map(b => b.id));
+                }
+                if (currentGridPos) {
+                    const objectsAtCurrentPosition = objectsAtDimensionGridPosition(
+                        calc,
+                        this.bot3D.dimension,
+                        currentGridPos
+                    );
+                    ids.push(...objectsAtCurrentPosition.map(b => b.id));
+                }
+
+                this.bot3D.dimensionGroup.simulation3D.ensureUpdate(ids);
+            }
             this._lastPos = currentGridPos;
+            this._lastSortOrder = currentSortOrder;
             this._lastHeight = currentHeight;
             this._nextRot = getBotRotation(
                 calc,
@@ -95,8 +157,8 @@ export class DimensionPositionDecorator extends AuxBot3DDecoratorBase {
             this._atPosition = false;
             this._atRotation = false;
             if (!this._lerp) {
-                this.bot3D.display.position.copy(this._nextPos);
-                this.bot3D.display.rotation.set(
+                this.bot3D.container.position.copy(this._nextPos);
+                this._rotationObj.rotation.set(
                     this._nextRot.x,
                     this._nextRot.z,
                     this._nextRot.y
@@ -122,24 +184,28 @@ export class DimensionPositionDecorator extends AuxBot3DDecoratorBase {
         return Math.abs(this._lastHeight - currentHeight) > 0.01;
     }
 
-    private _positionUpdated(currentGridPos: {
-        x: number;
-        y: number;
-        z: number;
-    }): boolean {
+    private _positionUpdated(
+        currentGridPos: {
+            x: number;
+            y: number;
+            z: number;
+        },
+        currentSortOrder: number
+    ): boolean {
         return (
             !this._lastPos ||
             (currentGridPos.x !== this._lastPos.x ||
                 currentGridPos.y !== this._lastPos.y ||
-                currentGridPos.z !== this._lastPos.z)
+                currentGridPos.z !== this._lastPos.z) ||
+            this._lastSortOrder !== currentSortOrder
         );
     }
 
     frameUpdate(calc: BotCalculationContext): void {
         if (this._lerp && this._nextPos && this._nextRot) {
             if (!this._atPosition) {
-                this.bot3D.display.position.lerp(this._nextPos, 0.1);
-                const distance = this.bot3D.display.position.distanceTo(
+                this.bot3D.container.position.lerp(this._nextPos, 0.1);
+                const distance = this.bot3D.container.position.distanceTo(
                     this._nextPos
                 );
                 this._atPosition = distance < 0.01;
@@ -153,14 +219,50 @@ export class DimensionPositionDecorator extends AuxBot3DDecoratorBase {
                     'XYZ'
                 );
                 const q = new Quaternion().setFromEuler(euler);
-                this.bot3D.display.quaternion.slerp(q, 0.1);
+                this._rotationObj.quaternion.slerp(q, 0.1);
 
-                const angle = this.bot3D.display.quaternion.angleTo(q);
+                const angle = this._rotationObj.quaternion.angleTo(q);
                 this._atRotation = angle < 0.1;
             }
 
             if (!this._atPosition || !this._atRotation) {
-                this.bot3D.display.updateMatrixWorld(true);
+                this.bot3D.container.updateMatrixWorld(true);
+            }
+        } else {
+            let update = false;
+            if (
+                ['billboard', 'billboardX', 'billboardZ'].indexOf(
+                    this._orientationMode
+                ) >= 0
+            ) {
+                const cameraRig = this.bot3D.dimensionGroup.simulation3D.getMainCameraRig();
+                const cameraWorld = new Vector3();
+                cameraWorld.setFromMatrixPosition(
+                    cameraRig.mainCamera.matrixWorld
+                );
+                this._rotationObj.lookAt(cameraWorld);
+                update = true;
+                if (this._orientationMode === 'billboardZ') {
+                    const euler = new Euler().setFromQuaternion(
+                        this._rotationObj.quaternion,
+                        'XYZ'
+                    );
+                    euler.y = 0;
+                    euler.z = 0;
+                    this._rotationObj.setRotationFromEuler(euler);
+                } else if (this._orientationMode === 'billboardX') {
+                    const euler = new Euler().setFromQuaternion(
+                        this._rotationObj.quaternion,
+                        'YXZ'
+                    );
+                    euler.x = 0;
+                    euler.z = 0;
+                    this._rotationObj.setRotationFromEuler(euler);
+                }
+            }
+
+            if (update) {
+                this.bot3D.updateMatrixWorld(true);
             }
         }
     }